feat: Migrate MyApplication page to TS
